requirements.txt
asgiref==3.8.1
Django==5.1.6
gunicorn==23.0.0
packaging==24.2
psycopg==3.2.4
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
djangorestframework==3.14.0
